+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Released on ??
 - Merged `File`, `Directory` and `Entry` into a unique struct called `File`. File types (symlink, file, directory) are now differentiated by the `file_type` attribute in `Metadata`.
 - `find` method is now optional, via the `find` feature (enabled by default)
 - Implemented `From` trait for `Metadata`.
+- `create` and `append` will now return a `WriteStream` instead of a box, which will contain the inner stream which supports `Write` and may support `Seek` (according to the protocol).
+- `read` will now return a `ReadStream` instead of a box, which will contain the inner stream which supports `Read` and may support `Seek` (according to the protocol).
 
 ## 0.1.1